requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# NLP and embeddings
sentence-transformers>=2.2.0
transformers>=4.20.0
torch>=1.9.0

# Visualization
matplotlib>=3.4.0

# Configuration
PyYAML>=6.0

# API requests
requests>=2.26.0

# Optional: for GPU support
# Install PyTorch with CUDA separately if needed:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118

